+++ .gitignore
@@ -3,3 +3,5 @@
 
 __pycache__
 .DS_Store
+
+./compiled.en.py
