--- compiled.en.py
@@ -1,9 +1,0 @@
-‫کچھ= 2‬
-
-‫if کچھ == 1:‬
-‫	print("سلام")‬
-‫else:‬
-‫	print("دنیا")‬
-
-‫ہندسہ =  120‬
-‫print(ہندسہ - int("۱۰۰۰"))‬
