+++ requirements.txt
@@ -1,0 +1,3 @@
+ply==3.11
+PyYAML==5.4.1
+Unidecode==1.3.2
