+++ samples/حساب۔پای
@@ -1,0 +1,13 @@
+ہندسہ =  ۱۲۰
+
+جمع = ہندسہ + عدد("۱۰۰۰")
+چھاپ(جمع)
+
+فرق = ۵۰۰۰ -  ۲۰۰۰
+چھاپ (فرق)
+
+ضرب = ۲۰ * ۳۰
+چھاپ (ضرب)
+
+تقسیم = ۱۰۰  / ۱۰
+چھاپ (تقسیم)
