+++ samples/مشروط۔پای
@@ -4,6 +4,3 @@
 	چھاپ("سلام")
 ورنہ:
 	چھاپ("دنیا")
-
-ہندسہ =  ۱۲۰
-چھاپ(ہندسہ - عدد("1000"))
