+++ ur_native.lang.yaml
@@ -7,6 +7,18 @@ numbers:
   end           :       "۹"
 
 reserved:
+  "۔"           :       "."
+  "،"           :       ","
+  "۱"           :       "1"
+  "۲"           :       "2"
+  "۳"           :       "3"
+  "۴"           :       "4"
+  "۵"           :       "5"
+  "۶"           :       "6"
+  "۷"           :       "7"
+  "۸"           :       "8"
+  "۹"           :       "9"
+  "۰"           :       "0"
   "چھاپ"        :       "print"
   "ورنہ اگر"    :       "elif"
   "اگر"         :       "if"
